move oauth-metadata out from route file
